requirements.txt
appdirs==1.4.3
asgiref==3.3.4
CacheControl==0.12.6
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2019.11.28
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cssselect2==0.3.0
defusedxml==0.6.0
distlib==0.3.0
distro==1.4.0
Django==3.2.3
django-background-tasks==1.2.5
django-compat==1.0.15
django-cors-headers==3.4.0
django-extensions==3.0.2
django-quill-editor==0.1.22
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
packaging==20.3
pdfkit==0.6.1
pep517==0.8.2
#Pillow==7.2.0
progress==1.5
#psycopg2==2.8.5
pycparser==2.20
PyJWT==1.7.1
pyparsing==2.4.6
Pyphen==0.9.5
pytoml==0.1.21
pytz==2020.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.3.1
tinycss2==1.0.2
urllib3==1.25.8
WeasyPrint==51
webencodings==0.5.1
wkhtmltopdf==0.2
